rename interp.py -> gp.py
